backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
curl_cffi==0.13.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
frozendict==2.4.7
gunicorn==25.1.0
idna==3.11
lxml==6.0.2
multitasking==0.0.12
numpy==2.4.3
packaging==26.0
pandas==3.0.1
peewee==4.0.2
platformdirs==4.9.4
protobuf==7.34.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vaderSentiment==3.3.2
waitress==3.0.2
websockets==16.0
yfinance==1.2.0
langchain==0.3.1
langchain-groq==0.2.1
langgraph==0.2.35
langchain-community==0.3.1
chromadb==0.5.11
